+++ BE/src/main/java/codesquad/shine/issuetracker/user/presentation/UserController.java
@@ -5,8 +5,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
+@RestController
 @RequestMapping("/api")
 public class UserController {
