+++ BE/src/main/java/codesquad/shine/support/auth/authentication/handler/AuthenticationSuccessHandler.java
@@ -1,6 +1,6 @@
 package codesquad.shine.support.auth.authentication.handler;
 
-import codesquad.shine.support.auth.authentication.context.Authentication;
+import codesquad.shine.support.auth.context.Authentication;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
