+++ BE/src/main/java/codesquad/shine/support/auth/authentication/provider/AuthenticationManager.java
@@ -1,7 +1,7 @@
 package codesquad.shine.support.auth.authentication.provider;
 
 import codesquad.shine.support.auth.authentication.AuthenticationException;
-import codesquad.shine.support.auth.authentication.context.Authentication;
+import codesquad.shine.support.auth.context.Authentication;
 
 public interface AuthenticationManager {
     Authentication authenticate(Authentication authentication) throws AuthenticationException;
