+++ BE/src/test/java/codesquad/shine/issuetracker/auth/AbstractAuthenticationProcessingFilterTest.java
@@ -4,12 +4,12 @@ import codesquad.shine.issuetracker.user.business.CustomUserDetailsService;
 import codesquad.shine.issuetracker.user.domain.User;
 import codesquad.shine.issuetracker.user.domain.UserRepository;
 import codesquad.shine.support.auth.authentication.AuthenticationException;
-import codesquad.shine.support.auth.authentication.context.Authentication;
 import codesquad.shine.support.auth.authentication.filter.UsernamePasswordAuthenticationFilter;
 import codesquad.shine.support.auth.authentication.handler.DefaultAuthenticationFailureHandler;
 import codesquad.shine.support.auth.authentication.handler.DefaultAuthenticationSuccessHandler;
 import codesquad.shine.support.auth.authentication.provider.AuthenticationManager;
 import codesquad.shine.support.auth.authentication.provider.DaoAuthenticationProvider;
+import codesquad.shine.support.auth.context.Authentication;
 import org.assertj.core.api.ThrowableAssert;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
